Split out ReactDOM setup for Reconciler injection
